move base stage to db
