+++ silkworm/node/stagedsync/execution_engine.hpp
@@ -28,8 +28,8 @@
 
 #include <silkworm/core/common/lru_cache.hpp>
 #include <silkworm/core/types/block.hpp>
+#include <silkworm/node/db/stage.hpp>
 #include <silkworm/node/stagedsync/execution_pipeline.hpp>
-#include <silkworm/node/stagedsync/stages/stage.hpp>
 
 #include "forks/extending_fork.hpp"
 #include "forks/main_chain.hpp"
