+++ silkworm/node/stagedsync/forks/canonical_chain.hpp
@@ -25,8 +25,8 @@
 #include <silkworm/core/common/lru_cache.hpp>
 #include <silkworm/core/types/block.hpp>
 #include <silkworm/node/db/access_layer.hpp>
+#include <silkworm/node/db/stage.hpp>
 #include <silkworm/node/stagedsync/execution_pipeline.hpp>
-#include <silkworm/node/stagedsync/stages/stage.hpp>
 
 namespace silkworm::stagedsync {
