+++ silkworm/node/stagedsync/forks/main_chain.hpp
@@ -27,8 +27,8 @@
 #include <silkworm/core/common/lru_cache.hpp>
 #include <silkworm/core/types/block.hpp>
 #include <silkworm/node/db/memory_mutation.hpp>
+#include <silkworm/node/db/stage.hpp>
 #include <silkworm/node/stagedsync/execution_pipeline.hpp>
-#include <silkworm/node/stagedsync/stages/stage.hpp>
 
 #include "canonical_chain.hpp"
 #include "verification_result.hpp"
