+++ silkworm/node/stagedsync/stages/stage_blockhashes.hpp
@@ -17,7 +17,7 @@
 #pragma once
 
 #include <silkworm/node/db/etl/collector_settings.hpp>
-#include <silkworm/node/stagedsync/stages/stage.hpp>
+#include <silkworm/node/db/stage.hpp>
 
 namespace silkworm::stagedsync {
