+++ silkworm/node/stagedsync/stages/stage_call_trace_index.hpp
@@ -21,7 +21,7 @@
 #include <silkworm/node/db/bitmap.hpp>
 #include <silkworm/node/db/etl/collector_settings.hpp>
 #include <silkworm/node/db/prune_mode.hpp>
-#include <silkworm/node/stagedsync/stages/stage.hpp>
+#include <silkworm/node/db/stage.hpp>
 
 namespace silkworm::stagedsync {
