+++ silkworm/node/stagedsync/stages/stage_finish.hpp
@@ -18,7 +18,7 @@
 
 #include <string>
 
-#include <silkworm/node/stagedsync/stages/stage.hpp>
+#include <silkworm/node/db/stage.hpp>
 
 namespace silkworm::stagedsync {
