+++ silkworm/node/stagedsync/stages/stage_headers.hpp
@@ -20,7 +20,7 @@
 #include <silkworm/core/types/hash.hpp>
 #include <silkworm/infra/common/measure.hpp>
 #include <silkworm/node/db/access_layer.hpp>
-#include <silkworm/node/stagedsync/stages/stage.hpp>
+#include <silkworm/node/db/stage.hpp>
 
 namespace silkworm::stagedsync {
