+++ silkworm/node/stagedsync/stages/stage_senders.hpp
@@ -32,7 +32,7 @@
 #include <silkworm/node/db/etl/collector.hpp>
 #include <silkworm/node/db/etl/collector_settings.hpp>
 #include <silkworm/node/db/prune_mode.hpp>
-#include <silkworm/node/stagedsync/stages/stage.hpp>
+#include <silkworm/node/db/stage.hpp>
 
 namespace silkworm::stagedsync {
