+++ silkworm/node/stagedsync/stages/stage_tx_lookup.hpp
@@ -19,7 +19,7 @@
 #include <silkworm/node/db/bitmap.hpp>
 #include <silkworm/node/db/etl/collector_settings.hpp>
 #include <silkworm/node/db/prune_mode.hpp>
-#include <silkworm/node/stagedsync/stages/stage.hpp>
+#include <silkworm/node/db/stage.hpp>
 
 namespace silkworm::stagedsync {
